code/src/Finsure VS code plugin/plugin/scripts/frontend/requirements.txt
streamlit
requests
google-generativeai
python-dotenv
fpdf